Changed wait times to be more dynamic
Removed all instances where "Sleep" was used
Replaced them with "Wait Until Element Is Visible", or "Wait Until Element Is Enabled" instead.
Also replaced the checks that verified that the element was present on the webpage with checks like those above.
Replaced all "magic numbers" used with for the time to wait with the variable standard_timeout